Fix deploy timeouts
Deploy is a streaming operation and we need the connection open
even with the presence of firewalls or load balancers that have
idle timeouts. As we don't control go-swagger autogenerated server
we send a "ping" to the client every 30 seconds by default (this
value may be changed using the TERESADEPLOY_KEEPALIVE_TIMEOUT env var).